Add config option to create to control whether a new option is selected or not
If the creation of the new option in production code is dependent on some other service--like inserting the new option in a database--there is a potential for that operation to fail.  When writing a unit test to verify how that failure is handled, it may not be desirable to select the new option given to `create` if the insert into the database failed.

This commit introduces a new option in the `CreateConfig` object called `autoSelect` that will enable or disable the automatic selection of the new option.

See issue #35 for more details.