Fix for autoSelect not implemented correctly
This is piggybacking off of issue #35 and pull request #36.

The implementation in #36 doesn't work as expected.  Regardless, I believe `await select(input, createOptionText, config);` should happen.  If I understand the code correctly, calling `await select(...)` is mimicking the user clicking `Create: "[whatever they entered]"`. I want the interaction of clicking create to happen, but when the create fails on the backend, I don't want this library to then try and find the new value (`findByText`) in the select menu because it shouldn't exist.